feat: Add nodejs-basics::modules::cjsToEsm task implementation
